+++ library/std/src/os/hermit/io/mod.rs
@@ -1,0 +1,10 @@
+#![stable(feature = "os_fd", since = "1.66.0")]
+
+mod owned;
+mod raw;
+
+// Export the types and traits for the public API.
+#[stable(feature = "os_fd", since = "1.66.0")]
+pub use owned::*;
+#[stable(feature = "os_fd", since = "1.66.0")]
+pub use raw::*;
